Move utils tests from test_panel -> test_utils
