+++ django_mongodb_extensions/debug_toolbar/panels/mql/forms.py
@@ -4,7 +4,6 @@ from bson import json_util
 from debug_toolbar.panels.sql.forms import SQLSelectForm
 from debug_toolbar.toolbar import DebugToolbar
 from django import forms
-from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import connections
 from django.utils.translation import gettext_lazy as _
