+++ pyproject.toml
@@ -17,7 +17,7 @@ test = [
 ]
 
 [tool.ruff.lint]
-select = ["I"]
+select = ["I", "F"]
 
 [tool.hatch.build.targets.wheel]
 packages = ["django_mongodb_extensions"]
